Show context menu at the correct location
